Move the release controller into its own package
This allows us to introduce new controllers over time without having to worry
about naming conflicts within a single package scope.
