+++ pkg/controller/controller.go
@@ -14,25 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+// Package controller contains Helm controllers
 package controller
-
-import (
-	ctrl "sigs.k8s.io/controller-runtime"
-
-	"github.com/crossplane-contrib/provider-helm/pkg/controller/release"
-
-	"github.com/crossplane/crossplane-runtime/pkg/logging"
-)
-
-// Setup creates all Helm controllers with the supplied logger and adds them
-// to the supplied manager.
-func Setup(mgr ctrl.Manager, l logging.Logger) error {
-	for _, setup := range []func(ctrl.Manager, logging.Logger) error{
-		release.Setup,
-	} {
-		if err := setup(mgr, l); err != nil {
-			return err
-		}
-	}
-	return nil
-}
