+++ pkg/controller/release/patch.go
@@ -1,4 +1,4 @@
-package controller
+package release
 
 import (
 	"context"
